+++ framework/hyper-express/index.ts
@@ -77,6 +77,7 @@ export type {
 } from '@forklaunch/hyper-express-fork';
 export type { Application } from './src/hyperExpressApplication';
 export type { Router } from './src/hyperExpressRouter';
+export type { ExpressOptions } from './src/types/hyperExpressOptions.types';
 
 export const handlers: {
   any: typeof any;