+++ framework/validator/src/zod/zodSchemaValidator.ts
@@ -251,6 +251,7 @@ export class ZodSchemaValidator
       format: 'binary',
       example: 'a utf-8 encodable string'
     });
+  type = <T>() => this.any as ZodType<T>;
 
   /**
    * Compiles schema if this exists, for optimal performance.