+++ framework/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @forklaunch/core
 
+## 0.9.10
+
+### Patch Changes
+
+- bump package versions, allow for validator custom types that resolve as any, export http framework options type
+- Updated dependencies
+  - @forklaunch/validator@0.6.10
+  - @forklaunch/common@0.3.10
+
 ## 0.9.9
 
 ### Patch Changes