+++ framework/infrastructure/S3/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @forklaunch/infrastructure-s3
 
+## 0.0.18
+
+### Patch Changes
+
+- bump package versions, allow for validator custom types that resolve as any, export http framework options type
+- Updated dependencies
+  - @forklaunch/common@0.3.10
+  - @forklaunch/core@0.9.10
+
 ## 0.0.17
 
 ### Patch Changes